+++ modules/tunnel.sh
@@ -1,37 +1,35 @@
 #!/bin/bash
 # guru-client tunneling functions 2021
 
-source common.sh
-#source system.sh
+declare -g tunnel_rc="/tmp/guru-cli_tunnel.rc"
+
 
 tunnel.help () {
-    # genereal help
-
-    gmsg -v1 -c white "guru-client tunnel help "
-    gmsg -v2
-    gmsg -v0 "usage:    $GURU_CALL tunnel status|ls|open|close|add|rm|start|end|install|remove] "
-    gmsg -v2
-    gmsg -v1 -c white  "commands:"
-    gmsg -v1 " ls               list, active highlighted"
-    gmsg -v1 " status           tunnel status "
-    gmsg -v1 " open <service>   open known tunnel set in user.cfg"
-    gmsg -v1 " close <service>  close open tunnels"
-    gmsg -v2 " add              add tunnel to .ssh/config"
-    gmsg -v2 " rm               remove tunnel from .ssh/config"
-    gmsg -v1 " install          install requirements "
-    gmsg -v1 " remove           remove installed requirements "
-    gmsg -v2 " poll start|end   start or end module status polling "
-    gmsg -v2
-    gmsg -v1 -c white  "example:"
-    gmsg -v1 "    $GURU_CALL tunnel open wiki"
-    gmsg -v2
+# tunnel commands
+    gr.msg -v1 "guru-client tunnel help " -c white
+    gr.msg -v2
+    gr.msg -v0 "usage:    $GURU_CALL tunnel status|ls|open|close|add|rm|start|end|install|remove] "
+    gr.msg -v2
+    gr.msg -v1 "commands:" -c white
+    gr.msg -v1 " ls               list, active highlighted"
+    gr.msg -v1 " status           tunnel status "
+    gr.msg -v1 " open <service>   open known tunnel set in user config"
+    gr.msg -v1 " close <service>  close open tunnels"
+    gr.msg -v2 " add              add tunnel to .ssh/config"
+    gr.msg -v2 " rm               remove tunnel from .ssh/config"
+    gr.msg -v1 " install          install requirements "
+    gr.msg -v1 " remove           remove installed requirements "
+    gr.msg -v2 " poll start|end   start or end module status polling "
+    gr.msg -v3 " hop              TBD hopping tunnel support"
+    gr.msg -v2
+    gr.msg -v1 "example:" -c white
+    gr.msg -v1 "    $GURU_CALL tunnel open wiki"
+    gr.msg -v2
 }
 
 
-# xprop -id 0x5600001
-
 tunnel.main () {
-    # command parser
+# command parser
 
     command="$1"; shift
     case "$command" in
@@ -64,38 +62,85 @@ tunnel.main () {
 }
 
 
+
+tunnel.rc () {
+# source configurations
+
+    if  [[ ! -f $tunnel_rc ]] || \
+        [[ $(( $(stat -c %Y $GURU_CFG/$GURU_USER/tunnel.cfg) - $(stat -c %Y $tunnel_rc) )) -gt 0 ]]
+        then
+            tunnel.make_rc && \
+                gr.msg -v1 -c dark_gray "$tunnel_rc updated"
+        fi
+
+    source $tunnel_rc
+}
+
+
+tunnel.make_rc () {
+# make core module rc file out of configuration file
+
+    if ! source config.sh ; then
+            gr.msg -c yellow "unable to load configuration module"
+            return 100
+        fi
+
+    if [[ -f $tunnel_rc ]] ; then
+            rm -f $tunnel_rc
+        fi
+
+    if ! config.make_rc "$GURU_CFG/$GURU_USER/tunnel.cfg" $tunnel_rc ; then
+            gr.msg -c yellow "configuration failed"
+            return 101
+        fi
+
+    chmod +x $tunnel_rc
+
+    if ! source $tunnel_rc ; then
+            gr.msg -c red "unable to source configuration"
+            return 202
+        fi
+}
+
+
 tunnel.status () {
-    # check tunnel is reachable
+# check tunnel is reachable
 
     # ptrintout timestamp and function name
-    gmsg -n -v1 -t "${FUNCNAME[0]}: "
+    gr.msg -n -v1 -t "${FUNCNAME[0]}: "
 
-    local tunnel_indicator_key='f'"$(daemon.poll_order tunnel)"
+    # check system is able to service
+    if [[ $GURU_TUNNEL_ENABLED ]] ; then
+            gr.msg -v1 -n -c green "enabled, " -k $GURU_TUNNEL_INDICATOR_KEY
+        else
+            gr.msg -v1 -c black "disabled" -k $GURU_TUNNEL_INDICATOR_KEY
+            return 1
+        fi
 
     # check system is able to service
     if [[ -f /usr/bin/ssh ]] ; then
-            gmsg -v1 -n -c green "available " -k $tunnel_indicator_key
+            gr.msg -v1 -n -c green "available " -k $GURU_TUNNEL_INDICATOR_KEY
         else
-            gmsg -v1 -c red "not installed" -k $tunnel_indicator_key
+            gr.msg -v1 -c red "not installed" -k $GURU_TUNNEL_INDICATOR_KEY
             return 2
         fi
 
     # list of tunnels all verbose levels
-    [[ $GURU_VERBOSE ]] && tunnel.ls
+    [[ $GURU_VERBOSE -gt 0 ]] && tunnel.ls
 
     # indicate user if active tunnels
-    if ps -x | grep -v grep | grep "ssh -L " | grep localhost >/dev/null; then
-            gmsg -n -v4 -c aqua "active tunnels" -k $tunnel_indicator_key
+    if ps -x | grep -v grep | grep "ssh -L " | grep -q localhost ; then
+            gr.msg -v3 -c aqua "active tunnels" -k $GURU_TUNNEL_INDICATOR_KEY
         else
-            gmsg -n -v4 -c reset "no active tunnels"
+            gr.msg -v3 -c green "no active tunnels" -k $GURU_TUNNEL_INDICATOR_KEY
         fi
-    echo
+
     return 0
 }
 
 
 tunnel.add () {
-    # add ssh tunnel to .ssh_config
+# add ssh tunnel to .ssh_config
 
     echo TBD
     # https://www.everythingcli.org/ssh-tunnelling-for-fun-and-profit-autossh/
@@ -114,21 +159,36 @@ tunnel.add () {
 
 
 tunnel.rm () {
-    # remove ssh tunnel from ssh config
+# remove ssh tunnel from ssh config
 
     echo TBD
     return 0
 }
 
 
+tunnel.check () {
+# check tunnel is active, if no input use first iten on list
+
+    local service=${GURU_TUNNEL_LIST[0]}
+    [[ $1 ]] && service=$1
+
+    tunnel.get_config $service
+
+    if ps -x | grep -v grep | grep "ssh -L $to_port:"  >/dev/null; then
+        return 0
+    else
+        return 1
+    fi
+}
+
+
 tunnel.toggle () {
-    # open default tunnel list and closes
+# open default tunnel list and closes
 
-    declare -l state="/tmp/tunnel.toggle"
+    local state="/tmp/tunnel.toggle"
 
-    if [[ -f $state ]] && tunnel.ls >/dev/null; then
-            tunnel.close # &&
-            rm $state
+    if [[ -f $state ]] && tunnel.check >/dev/null; then
+            tunnel.close && rm $state
         else
             tunnel.open ${GURU_TUNNEL_DEFAULT[@]} # &&
             touch $state
@@ -139,84 +199,63 @@ tunnel.toggle () {
 
 
 tunnel.ls () {
-    # list of ssh tunnels
+# list of ssh tunnels
 
     local all_services=(${GURU_TUNNEL_LIST[@]})
-    [[ $1 ]] && all_services=($@)
-    local tunnel_indicator_key='f'"$(daemon.poll_order tunnel)"
+    # [[ $1 ]] && all_services=($@)
     local _return=1
 
-    # non color system get only list og active tunnels and do not set corsair stuff
-    if ! [[ $DISPLAY ]] ; then
-        for service in ${all_services[@]} ; do
-            tunnel.get_config $service
-            if ps -x | grep -v grep | grep "ssh" | grep "$to_port:" >/dev/null; then
-                    gmsg "$service"
-                    _return=0
-                fi
-            done
-
-        return $_return
-        fi
-
-    # check only given service
-    if [[ $service ]] ; then
-            tunnel.get_config $service
-            if ps -x | grep -v grep | grep "ssh -L $to_port:localhost:" >/dev/null; then
-                    gmsg -v2 -c aqua "$service"  -k $tunnel_indicator_key
-                    return 0
-                else
-                    gmsg -v3 -c green "$service" -k $tunnel_indicator_key
-                    return 1
-                fi
-        fi
-
-    # check all tunnels
-    gmsg -n -v2 "tunnels: "
+    gr.msg -n -v2 "tunnels: "
     for service in ${all_services[@]} ; do
         tunnel.get_config $service
         if ps -x | grep -v grep | grep "ssh" | grep "$to_port:" >/dev/null; then
-                gmsg -n -c aqua "$service " -k $tunnel_indicator_key
+                gr.msg -n -c aqua "$service "
                 _return=0
             else
-                gmsg -n -c light_blue "$service "
+                gr.msg -n -c dark_cyan "$service "
             fi
         done
-    echo
+    gr.msg
+}
 
-    if (( _return > 0 )) ; then
-        gmsg -v3 -c green "reset" -k $tunnel_indicator_key
-        return 1
-    else
-        return 0
-    fi
+
+tunnel.in_list () {
+# dub
+    local service_name=$1
+
+    for service_list_item in ${GURU_TUNNEL_LIST[@]} ; do
+            [[ ${service_list_item} == $service_name ]] && return 0
+        done
+    return 1
 }
 
 
 tunnel.get_config () {
-    # get tunnel configuration an populate common variables
+# get tunnel configuration an populate common variables
 
-    declare -l service_name=$1
-    declare -l all_services=(${GURU_TUNNEL_LIST[@]})
-    declare -l options=${GURU_TUNNEL_OPTIONS[@]}
+    local i=0
+    local service_nr=0
+    local value=
+    local service_name=$1
+    local all_services=(${GURU_TUNNEL_LIST[@]})
+    local options=${GURU_TUNNEL_OPTIONS[@]}
 
-    # exit is not in list
-    if ! echo "${all_services[@]}" | grep "$service_name" >/dev/null ; then
-        gmsg -c yellow "no service $service_name in user configuration"
+    # exit if not in list
+    if ! tunnel.in_list $service_name ; then
+        gr.msg -c yellow "no service $service_name in user configuration"
         return 1
         fi
 
     # get location in array
-    local service_nr=0
     for service in ${all_services[@]} ; do
             if [[ "$service" == "$service_name" ]] ; then break ; fi
             (( service_nr++ ))
         done
 
-    local i=0
+    # evaluate variables
     for parameter in ${options[@]} ; do
-        local value="GURU_TUNNEL_${GURU_TUNNEL_LIST[$service_nr]^^}[$i]"
-        #gmsg -c pink "$value $parameter=${!value}"
+        value="GURU_TUNNEL_${GURU_TUNNEL_LIST[$service_nr]^^}[$i]"
+        gr.msg -v4 -c pink "$value $parameter=${!value}"
         eval $parameter=${!value}
         (( i++ ))
         done
@@ -225,14 +264,14 @@ tunnel.get_config () {
 
 
 tunnel.open () {
-    # open local ssh tunnel
+# open local ssh tunnel
 
     local service_name=()
 
     if ! [[ $DISPLAY ]] ; then
-            gmsg -v1 "Seems that session is not local, least it does not have DISPLAY variable set."
-            gmsg -v2 "Therefore multible terminal windows cannot be lauched automatically."
-            gmsg -v0 "Open another terminal and paste following commands to avoid tunnel to open here."
+            gr.msg -v1 "Seems that session is not local, least it does not have DISPLAY variable set."
+            gr.msg -v2 "Therefore multiple terminal windows cannot be launch automatically."
+            gr.msg -v0 "Open another terminal and paste following commands to avoid tunnel to open here."
         fi
 
     if [[ $1 ]] ; then
@@ -241,71 +280,42 @@ tunnel.open () {
             service_name=(${GURU_TUNNEL_DEFAULT[@]})
         fi
 
-    local tunnel_indicator_key='f'"$(daemon.poll_order tunnel)"
     local ssh_param="-o ServerAliveInterval=15"
 
     for _service in ${service_name[@]} ; do
 
             if ! tunnel.get_config $_service ; then
-                gindicate error $tunnel_indicator_key
+                gr.ind error $GURU_TUNNEL_INDICATOR_KEY
                 continue
             fi
 
             local url="http://localhost:$to_port"
             [[ $url_end ]] && url="$url/$url_end"
 
-            if tunnel.ls $_service ; then
-                    # gmsg -v2 -n -c light_blue "$_service "
-                    gmsg -v1 -c white "$_service: $url"
+            if tunnel.check $_service ; then
+                    # gr.msg -v2 -n -c light_blue "$_service "
+                    gr.msg -v1 -c white "$_service: $url"
                     continue
                 fi
 
             if [[ $DISPLAY ]] ; then
-
-                    gnome-terminal  --hide-menubar --geometry 47x10 --zoom 0.5 \
-                                    --title "$_service"  -- \
+                    # open terminal and make ssh tunnel
+                    # TBD terminal title changes to USER@HOST and is too much of effort to keep as $_service
+                    gnome-terminal  --hide-menubar --geometry 45x12 --zoom 0.5 \
+                                    --title "$_service" -- \
                                     ssh -L $to_port:localhost:$from_port \
                                     $user@$domain -p $ssh_port \
                                     $ssh_param \
                                     ; pidof gnome-terminal &
 
-                    # echo $_output
-                    # TBD not able to get pig of gnome-terminal session =/
-                    # local process_id=$(sh -c "echo && ; gnome-terminal --hide-menubar --geometry 47x10 --zoom 0.5 --title $_service  -- ssh -L $to_port:localhost:$from_port $user@$domain -p $ssh_port")
-                    # process_id=$(sh -c 'gnome-terminal --hide-menubar --geometry 47x10 --zoom 0.5 --title kukka  -- ssh -L 8181:localhost:8181 casa@roima -p 22')
-                    # process_id=$(sh -c 'echo $$; exec gnome-terminal --hide-menubar --geometry 47x10 --zoom 0.5 --title kukka  -- ssh -L 8181:localhost:8181 casa@roima -p 22')
-                    #gmsg -v2 -c white -n "$_service $from_port "
-                    # check if graphical environment or forced not to open window
-                    # local process_id=$!
-                    # gmsg -v2 -c white -n "pid: $process_id "
-
-                    # get tunnel pid
-                    # local tunnel_pid=$(ps a | \
-                    #         grep -v grep | \
-                    #         grep "ssh -L" | \
-                    #         grep $from_port | \
-                    #         head -n 1  | sed 's/^[[:space:]]*//' | \
-                    #         cut -d " "  -f 1 \
-                    #     )
-                    # # gmsg -v2 -c white -n "tunnel_process: $tunnel_pid "
-
-                    # try to minimaze window
-                    #local window_id=$(system.get_window_id $tunnel_pid)
-                    # local window_id=$(system.get_window_id $process_id)
-
-                    # if [[ $window_ssh id ]] ; then
-                    #         # gmsg -v2 -c white -n "window_id: $window_id "
-                    #         xdotool windowminimize ${window_id}
-                    #     fi
-
                 else
                     # console environment
-                    gmsg -v0 -c light_blue "ssh -L $to_port:localhost:$from_port $user@$domain -p $ssh_port"
-                    #gindicate fail $tunnel_indicator_key
+                    gr.msg -v0 -c dark_cyan "ssh -L $to_port:localhost:$from_port $user@$domain -p $ssh_port"
+                    #gr.ind fail $GURU_TUNNEL_INDICATOR_KEY
 
                 fi
 
-            gmsg -v1 -c aqua "$_service: $url" -k $tunnel_indicator_key
+            gr.msg -v1 -c aqua "$_service: $url" -k $GURU_TUNNEL_INDICATOR_KEY
 
         done
     return 0
@@ -313,10 +323,9 @@ tunnel.open () {
 
 
 tunnel.close () {
-    # close ssh tunnel
+# close ssh tunnel
 
     local data_folder="$GURU_SYSTEM_MOUNT/tunnel/$(hostname)"
-    local tunnel_indicator_key='f'"$(daemon.poll_order tunnel)"
     local pid_list=()
 
     if [[ $1 ]] ; then
@@ -347,40 +356,68 @@ tunnel.close () {
 
     # check that pid is number
     if ! [[ ${pid_list[@]} ]] ; then
-        gmsg -v1 "no tunnels"
+        gr.msg -v1 "no tunnels"
         return 0
     fi
 
     # kill those tunnels
     for pid in ${pid_list[@]} ; do
 
-            gmsg -v2 "killing pid $pid"
+            gr.msg -v2 "killing pid $pid"
             if ! echo $pid | grep -E ^\-?[0-9]?\.?[0-9]+$ >/dev/null; then
-                    gmsg -v2 -c yellow "tunnel '$service_name' not found "
+                    gr.msg -v2 -c yellow "tunnel '$service_name' not found "
                     return 0
                 fi
 
             if kill -15 $pid ; then
-                    gmsg -v1 -c green "$pid killed"
-                    gindicate ok $tunnel_indicator_key
+                    gr.msg -v1 -c green "$pid killed"
+                    gr.ind ok $GURU_TUNNEL_INDICATOR_KEY
                 else
-                    kill -9 $pid || gmsg -c yellow "$pid kill failed"
+                    kill -9 $pid || gr.msg -c yellow "$pid kill failed"
                 fi
         done
 
-    if tunnel.ls >/dev/null; then
-            gmsg -v2 -c yellow "active tunnels detected"
-            gindicate error $tunnel_indicator_key
+    if tunnel.check >/dev/null; then
+            gr.msg -v2 -c yellow "active tunnels detected"
+            gr.ind error $GURU_TUNNEL_INDICATOR_KEY
         fi
 
     return 0
 }
 
 
-tunnel.tmux () {
-    # add new session "tunnels", open new pane to main window for all tunnel sessions
 
-    gmsg -v2 -c blue "TBD tmux ahve A lot of vibstagels, later"
+ssh.end_remote_tunnel_sessions () {
+
+    local _server="$GURU_USER@$GURU_ACCESS_DOMAIN"
+    local _port="$GURU_ACCESS_PORT"
+
+    [[ $1 ]] && _server=$1
+    [[ $2 ]] && _port=$2
+
+    local _ifs=$IFS ; IFS=$'\n'
+    local _pid_list=($(\
+        ssh $_server -p $_port -- ps -xf  \
+            | grep -v grep \
+            | grep '?' \
+            | grep sshd \
+            | grep "$GURU_USER@notty"\
+            | sed -e's/  */ /g' \
+            | cut -d' ' -f 2
+            ))
+    IFS=$_ifs
+
+    #for item in ${_pid_list[@]} ; do
+            echo ssh $_server -p $_port -- kill ${_pid_list[@]}
+     #   done
+}
+
+
+
+tunnel.tmux () {
+# add new session "tunnels", open new pane to main window for all tunnel sessions
+
+    gr.msg -v2 -c blue "TBD tmux ahve A lot of vibstagels, later"
     # local session=tunnels
     # local window=${session}:0
     # local pane=${window}.0
@@ -388,9 +425,9 @@ tunnel.tmux () {
     # if ! tmux attach -t $session ; then
     #         gnome-terminal -- tmux new -s $session &
 
-    #         sleep 1 ; gmsg "sending C-b"
+    #         sleep 1 ; gr.msg "sending C-b"
     #         tmux send-keys -t "$pane" C-b C-b
-    #         sleep 0,2 : gmsg "sending d"
+    #         sleep 0,2 : gr.msg "sending d"
     #         tmux send-keys -t "$pane" d
     #     fi
 
@@ -402,17 +439,16 @@ tunnel.tmux () {
 
 
 tunnel.poll () {
-    # daemon poller will run this
+# daemon poller will run this
 
     local _cmd="$1" ; shift
-    tunnel_indicator_key='f'"$(daemon.poll_order tunnel)"
 
     case $_cmd in
         start )
-            gmsg -v1 -t -c black "${FUNCNAME[0]}: tunnel status polling started" -k $tunnel_indicator_key
+            gr.msg -v1 -t -c black "${FUNCNAME[0]}: tunnel status polling started" -k $GURU_TUNNEL_INDICATOR_KEY
             ;;
         end )
-            gmsg -v1 -t -c reset "${FUNCNAME[0]}: tunnel status polling ended" -k $tunnel_indicator_key
+            gr.msg -v1 -t -c reset "${FUNCNAME[0]}: tunnel status polling ended" -k $GURU_TUNNEL_INDICATOR_KEY
             ;;
         status )
             tunnel.status $@
@@ -425,18 +461,19 @@ tunnel.poll () {
 
 
 tunnel.requirements () {
-    # install and remove needed applications. input "install" or "remove"
+# install and remove needed applications. input "install" or "remove"
 
     local action=$1
     [[ "$action" ]] || read -r -p "install or remove? :" action
     local require="ssh rsync"
-    gmsg -c yellow "need to install $require, ctrl+c? or input local "
-    sudo apt update && eval sudo apt "$action" "$require" && gmsg -c yellow "guru is now ready to tunnel"
+    gr.msg -c yellow "need to install $require, ctrl+c? or input local "
+    sudo apt update && eval sudo apt "$action" "$require" && gr.msg -c white "guru is now ready to tunnel"
 }
 
+tunnel.rc
 
 if [[ "${BASH_SOURCE[0]}" == "${0}" ]]; then
-    source $GURU_RC
+    # source $GURU_RC
     tunnel.main "$@"
     exit 0
 fi
